Adds the Disabler Revolver and Laser Shotgun
Every Head of Staff who previously did not have a weapon now has a Disabler Revolver. It's basically a weaker disabler with 2/3 of the stamina damage and only 6-7 shots.

Also, security know has a Laser Shotgun in the Armory. It's a semi-powerful weapon, with a wide hit area, but only has 2 shots. 
Still gonna need to see if it needs to be buffed or nerfed at a later time, depending on player usage.